Added basic nix support via a flake
This commit utilises some cuda and python packages from nixpkgs
to fetch dependencies, and also provides missing dependencies
and upgrades to several dependencies. It allows one to run
`nix build .#pytket-cutensornet` (or `nix build` for short) to build
the tket extension.

Tests can't be run within the normal `nix flake check` context
as the GPU drivers aren't available within the sandbox, so tests
can be run with `nix run .#tests` instead.

`nix develop` within the source root can be used to enter a shell
with pytket's cutensornet backend available. If and when this commit
is merged into main, users will be able to run:

`nix develop github:cqcl/pytket-cutensornet`

and have the environment fully loaded for their use.

Cuda support is verified on nixos. Verification on non-nixos nix
installations will be required for general use. Integration with
Cachix will be extremely beneficial due to the size and duration
of the builds.