Feature: MPS support for gates on non-adjacent qubits (#119)
# Description

MPS algorithms ``MPSxGate`` and ``MPSxMPO`` now support simulation of
two-qubit gates acting on non-adjacent qubits.

# Related issues

It relates to #107. As discussed in the comments on that issue,
sometimes applying `prepare_circuit_mps` to add SWAPs so that all gates
act on adjacent qubits is advantageous for runtime, hence, it is still
supported. However, it is generally advised to skip this preprocessing.

# Checklist

- [x] I have run the tests on a device with GPUs.
- [x] I have performed a self-review of my code.
- [x] I have commented hard-to-understand parts of my code.
- [x] I have made corresponding changes to the public API documentation.
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have updated the changelog with any user-facing changes.