Restore old docs and workflow (#140)
# Description

There's some problems with the new version of docs (listed in the linked
issue). As such, for now `pytket-cutensornet` will continue maintaining
the old version of the docs, and using it as the main version. Once
these issues are solved, I'm happy to make another release switching to
the new docs.

# Related issues

#139