docs: use `pytket-docs-theming` submodule (#153)
# Description

Updating the docs build to use the
[pytket-docs-theming](https://github.com/CQCL/pytket-docs-theming)
repository as a submodule. This means the docs can be built locally with
the latest theme config.

I also took great pleasure in deleting a bunch of legacy docs files ;).
If we go with this solution then we don't these anymore.

Take a look at the README for building the docs locally ->
https://github.com/CQCL/pytket-docs-theming/tree/main/extensions#building-the-api-docs-for-pytket-extensions

I've added some bash scripts to automate installing dependencies and
building the docs. Note that these scripts are only intended for local
development and will not be used in the website build. I don't want to
have to make sure that they match in all of the different repositories.

<img width="1421" alt="Screenshot 2024-09-04 at 01 27 54"
src="https://github.com/user-attachments/assets/8da592c8-8a03-495b-be54-7f24c018e2a1">


TODOS

- [x] Finalise how docs dependencies are specified. (proposal added in
README) Happy to have input on this
- [x] Fix docs build in C.I.
- [x] Discuss whether this is feasible for all other extenions that we
care about.


# Related issues

Please mention any github issues addressed by this PR.